remove useless exceptions; move model stub and base model to own file; move sequence diagrams to own folder in docs
