+++ docs/sequence_diagrams/Дообучение модели - последовательность.md
