+++ docs/sequence_diagrams/Загрузка новых данных - последовательность.md
