+++ docs/sequence_diagrams/Запрос прогноза - последовательность.md
