+++ docs/Схема БД.md
@@ -2,53 +2,37 @@
 
 ```mermaid
 erDiagram
-    GEO_CACHE {
-        string city_name PK
+    CITY {
+        int id PK
+        string name "уникальное"
         float latitude
         float longitude
-        timestamp cached_at
     }
 
-    WEATHER_OBSERVATIONS {
-        uuid id PK
-        string city_name FK
-        date obs_date
-        float temperature_mean
-        timestamp ingested_at
+    WEATHER_DAILY {
+        int id PK
+        int city_id FK
+        date date
+        float temperature_2m_mean
+        float temperature_2m_min
+        float temperature_2m_max
+        float apparent_temperature_mean
+        float apparent_temperature_min
+        float apparent_temperature_max
+        float precipitation_sum
+        float rain_sum
+        float snowfall_sum
+        float precipitation_hours
+        float pressure_msl_mean
+        float relative_humidity_2m_mean
+        float wind_speed_10m_max
+        float wind_gusts_10m_max
+        float shortwave_radiation_sum
+        float sunshine_duration
+        float uv_index_max
+        float daylight_duration
+        int weather_code
     }
 
-    FORECAST_LOG {
-        uuid id PK
-        string city_name FK
-        date forecast_date
-        float predicted_temp
-        float actual_temp "nullable до факта"
-        float error "nullable"
-        timestamp created_at
-    }
-
-    MAE_LOG {
-        uuid id PK
-        string city_name FK
-        date window_end
-        float mae_7d
-        float baseline_mae_7d
-        boolean threshold_exceeded
-        timestamp calculated_at
-    }
-
-    MODEL_VERSIONS {
-        uuid id PK
-        string model_type
-        string artifact_path
-        float test_mae
-        float baseline_mae
-        boolean is_active
-        boolean failed
-        timestamp trained_at
-    }
-
-    GEO_CACHE ||--o{ WEATHER_OBSERVATIONS : "city"
-    GEO_CACHE ||--o{ FORECAST_LOG : "city"
-    GEO_CACHE ||--o{ MAE_LOG : "city"
+    CITY ||--o{ WEATHER_DAILY : "has"
 ```
